==> manuscript/brain/non_normalized_code/04_DGE_overlap_25hr.R <==
#social transitions withcontrols. 

# libraries 
library(limma)
library(edgeR)
library(Mus.musculus)
organism = 'org.Mm.eg.db'
library(organism, character.only = TRUE)
library(biomaRt)
library(AnnotationDbi)
library(pheatmap)
library(annotables)
library(clusterProfiler)
library(enrichplot)
library(organism, character.only = TRUE)
library(DOSE)
library(tidyverse)
grcm38 # mouse genes


#ALL GROUPS 
my_logFC_threshold = 0.2

# outlier removed 

limma_list<- readRDS("manuscript/brain/results/limma_PFC_ReorganizedGroups_outlierremoved.RDS") %>% 
  map(~distinct(.)) %>% 
  map(~filter(.,abs(logFC) >= my_logFC_threshold)) %>%
  map(~filter(.,P.Value <0.05)) %>% 
  map(~ left_join(., grcm38 %>% dplyr::select(symbol, entrez))) %>% 
  map(~filter(.,!is.na(entrez))) 

adom <- limma_list$ascdom

doms <- limma_list$domsub

dd <- limma_list$desdom

da <- limma_list$desasc

as <- limma_list$ascsub

dsub <- limma_list$dessub

xx <- adom %>% rbind(dd,as,dsub)

all <- unique(xx$entrez)


#upreg
adom_up <- adom %>% filter(logFC >= 0.2) %>% arrange(-logFC)
doms_up <- doms %>% filter(logFC >= 0.2)%>% arrange(-logFC)
dd_up <- dd %>% filter(logFC >= 0.2)%>% arrange(-logFC) 
da_up <- da %>% filter(logFC >= 0.2)%>% arrange(-logFC)
as_up <- as %>% filter(logFC >= 0.2)%>% arrange(-logFC)
dsub_up <- dsub %>% filter(logFC >= 0.2)%>% arrange(-logFC)


#downreg
adom_down <- adom %>% filter(logFC <= -0.2)%>% arrange(logFC)
doms_down <- doms %>% filter(logFC <= -0.2)%>% arrange(logFC)
dd_down <- dd %>% filter(logFC <= -0.2)%>% arrange(logFC)
da_down <- da %>% filter(logFC <= -0.2)%>% arrange(logFC)
as_down <- as %>% filter(logFC <= -0.2)%>% arrange(logFC)
dsub_down <- dsub %>% filter(logFC <= -0.2)%>% arrange(logFC)




# start with TRN groups with dom
# upsetter Plot 
library(UpSetR)
library(workflowr)
library(ComplexUpset)
listInput <- list(adom_up= adom_up$symbol, dd_up=dd_up$symbol, adom_down = adom_down$symbol, 
                  dd_down = dd_down$symbol)


UpSetR::upset(fromList(listInput), nsets = 4, order.by = "freq", keep.order = F)

#getting overlap genes with for dom and trn groups 
dt_up25 <- adom_up$symbol[adom_up$symbol %in% dd_up$symbol] %>% as.data.frame()
dt_down25 <- adom_down$symbol[adom_down$symbol %in% dd_down$symbol]%>% as.data.frame()


#significant test
mat <- matrix(c(27,0,0,33),ncol=2)
# Perform chi-squared test
chi_square <- chisq.test(mat)


# subs
listInput <- list(dsub_up= dsub_up$symbol, as_up=as_up$symbol, dsub_down = dsub_down$symbol, 
                  as_down = as_down$symbol)


UpSetR::upset(fromList(listInput), nsets = 4, order.by = "freq", keep.order = F)


#getting overlap genes 
st_up25 <- as_up$symbol[as_up$symbol %in% dsub_up$symbol] %>% as.data.frame()
st_down25 <- as_down$symbol[as_down$symbol %in% dsub_down$symbol]%>% as.data.frame()

#significant test
mat <- matrix(c(98,0,0,85),ncol=2)
# Perform chi-squared test
chi_square <- chisq.test(mat)



# looking for overlap in transition groups between sub and dom comparision 
listInput <- list(st_up25= st_up25$., dt_up25=dt_up25$., st_down25 = st_down25$.,dt_down25 = dt_down25$.)


UpSetR::upset(fromList(listInput), nsets = 4, order.by = "freq", keep.order = F)

#getting overlap genes 
t_up25 <- st_up25$.[st_up25$. %in% dt_up25$.] %>% as.data.frame()
t_down25 <- st_down25$.[st_down25$. %in% dt_down25$.]%>% as.data.frame()

#significant test
mat <- matrix(c(8,0,0,11),ncol=2)
# Perform chi-squared test
chi_square <- chisq.test(mat)



colnames(dt_up25)[1]<- "symbol"
dt_up <- dt_up25 %>% mutate(regulated = "up") %>% mutate(condition = "TRN vs DOM") %>% mutate(time = 25)

colnames(dt_down25)[1]<- "symbol"
dt_down <- dt_down25 %>% mutate(regulated = "down") %>% mutate(condition = "TRN vs DOM") %>% mutate(time =25 )

dt <- dt_up %>%  rbind(dt_down)

df_dd <- dd %>% select(symbol, dd_logFC= logFC, dd_pv= P.Value) %>%
  mutate(time = 25) %>%  filter(symbol %in% dt$symbol) %>% filter(symbol != "")
df_adom <- adom %>% select(symbol, adom_logFC= logFC, adom_pv= P.Value) %>%
  mutate(time = 25) %>% filter(symbol %in% dt$symbol)%>% filter(symbol != "")

dd_dt <- df_dd %>% full_join(df_adom)

dd_dt25 <- dd_dt %>% full_join(dt) %>% na.omit(. )



colnames(st_up25)[1]<- "symbol"
st_up <- st_up25 %>% mutate(regulated = "up") %>% mutate(condition = "TRN vs SUB") %>% mutate(time = 25)

colnames(st_down25)[1]<- "symbol"
st_down <- st_down25 %>% mutate(regulated = "down") %>% mutate(condition = "TRN vs SUB") %>% mutate(time = 25)

st <- st_up %>%  rbind(st_down)

df_as <- as %>% select(symbol, as_logFC= logFC, as_pv= P.Value) %>%
  mutate(time = 25) %>%  filter(symbol %in% st$symbol)
df_ds<- dsub %>% select(symbol, ds_logFC= logFC, ds_pv= P.Value) %>%
  mutate(time = 25) %>% filter(symbol %in% st$symbol)

dd_st <- df_as %>% full_join(df_ds)

dd_st25 <- dd_st %>% full_join(st)



# source("manuscript/brain/04DEG-overlap_70min.R")
dt_all <- dd_dt70 %>% rbind(dd_dt25)
write.csv(dt_all,"manuscript/brain/results_tables/TRNgenesvsDOM_bothtimepoints.csv", row.names = F)

st_all <- dd_st70 %>% rbind(dd_st25)
write.csv(st_all,"manuscript/brain/results_tables/TRNgenesvsSUB_bothtimepoints.csv", row.names = F)




source("manuscript/brain/04_DGE_overlap_70min.R")
# dominant vs trn in both time points 
listInput <- list(dt_up= dt_up$., dt_up25=dt_up25$., dt_down = dt_down$.,dt_down25 = dt_down25$.)


UpSetR::upset(fromList(listInput), nsets = 4, order.by = "freq", keep.order = F)


dt_up[dt_up$. %in% dt_up25$.,]
dt_down[dt_down$. %in% dt_down25$.,]

# subordinates vs trn in both time points 
listInput <- list(st_up= st_up$., st_up25=st_up25$., st_down = st_down$.,st_down25 = st_down25$.)


UpSetR::upset(fromList(listInput), nsets = 4, order.by = "freq", keep.order = F)

st_up[st_up$. %in% st_up25$.,]
st_down[st_down$. %in% st_down25$.,]


st_down[st_down$. %in% st_up25$.,]
